+++ src/test/java/io/github/prayag/saucedemo/tests/data/SauceDemoUser.java
@@ -5,6 +5,7 @@ public enum SauceDemoUser {
   STANDARD("standard_user"),
   LOCKED_OUT("locked_out_user"),
   PROBLEM("problem_user"),
+  ERROR("error_user"),
   PERFORMANCE_GLITCH("performance_glitch_user");
 
   private final String username;