+++ src/test/java/io/github/prayag/saucedemo/tests/data/TestGroups.java
@@ -12,6 +12,6 @@ public final class TestGroups {
   public static final String CHECKOUT = "checkout";
   public static final String JOURNEY = "journey";
   public static final String PERSONA = "persona";
-  public static final String FRAMEWORK = "framework";
   public static final String ACCESSIBILITY = "accessibility";
+  public static final String VISUAL = "visual";
 }